# lib/units/api/swagger/api_v1_generated.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.3.0",
    "title": "Smartphone Test Farm",
    "description": "Control and manages real Smartphone devices from browser and restful apis",
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "STF Team",
      "email": "contact@openstf.io",
      "url": "http://openstf.io/"
    }
  },
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/octet-stream"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
    "/devicelog/getlogs": {
      "get": {
        "summary": "Get getlogs",
        "description": "Get All Projects with conditon",
        "operationId": "getLogs",
        "tags": [
          "devicelog"
        ],
        "parameters": [
          {
            "name": "startdate",
            "in": "query",
            "description": "Projects startdate",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get DeviceLogs",
            "schema": {
              "$ref": "#/definitions/DeviceLogResponse"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
    "/devicelog/getStatisticsPerGroup": {
      "get": {
        "summary": "Get getStatisticsPerGroup",
        "description": "Get All Projects with conditon",
        "operationId": "getStatisticsPerGroup",
        "tags": [
          "devicelog"
        ],
        "parameters": [
          {
            "name": "startdate",
            "in": "query",
            "description": "Projects getStatisticsPerGroup",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get getStatisticsPerGroup",
            "schema": {
              "$ref": "#/definitions/DeviceLogPerGroup"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
    "/devicelog/getStatisticsPerDate": {
      "get": {
        "summary": "Get getStatisticsPerDate",
        "description": "Get All Projects with conditon",
        "operationId": "getStatisticsPerDate",
        "tags": [
          "devicelog"
        ],
        "parameters": [
          {
            "name": "startdate",
            "in": "query",
            "description": "Projects getStatisticsPerDate",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get getStatisticsPerDate",
            "schema": {
              "$ref": "#/definitions/DeviceLogPerDate"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
    "/devicelog/getStatisticsPerCustom": {
      "get": {
        "summary": "Get getStatisticsPerCustom",
        "description": "Get All Projects with conditon",
        "operationId": "getStatisticsPerCustom",
        "tags": [
          "devicelog"
        ],
        "parameters": [
          {
            "name": "startdate",
            "in": "query",
            "description": "Projects getStatisticsPerCustom",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get getStatisticsPerCustom",
            "schema": {
              "$ref": "#/definitions/DeviceLogPerCustom"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "accessTokenAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}